% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerR-package.R
\docType{package}
\name{cancerR-package}
\alias{cancerR}
\alias{cancerR-package}
\title{cancerR: Classification of Cancer Using Administrative Data}
\description{
Classifies the type of cancer using routinely collected data commonly found in cancer registries from pathology reports. The package implements the International Classification of Diseases for Oncology, 3rd Edition site (topography), histology (morphology), and behaviour codes of neoplasms to classify cancer type \url{https://www.who.int/standards/classifications/other-classifications/international-classification-of-diseases-for-oncology}. Classification in children utilize the International Classification of Childhood Cancer by Steliarova-Foucher et al. (2005) \doi{10.1002/cncr.20910}. Adolescent and young adult cancer classification is based on Barr et al. (2020) \doi{10.1002/cncr.33041}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/giancarlodigi/cancerR}
  \item Report bugs at \url{https://github.com/giancarlodigi/cancerR/issues}
}

}
\author{
\strong{Maintainer}: Giancarlo Di Giuseppe \email{digi.giancarlo@proton.me} (\href{https://orcid.org/0000-0001-5112-7560}{ORCID}) [copyright holder]

}
\keyword{internal}
