% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerR-package.R
\docType{package}
\name{cancerR-package}
\alias{cancerR}
\alias{cancerR-package}
\title{cancerR: Classification of Cancer Using Administrative Data}
\description{
Classifies the type of cancer using routinely collected data commonly found in cancer registries from pathology reports. The package implements the International Classification of Diseases for Oncology, 3rd Edition (ICD-O-3) and the World Health Organization (WHO) classification of tumors. Information on the site (topography), histology (morphology), and behaviour of neoplasms are used to classify the type of cancer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/giancarlodigi/cancerR}
  \item Report bugs at \url{https://github.com/giancarlodigi/cancerR/issues}
}

}
\author{
\strong{Maintainer}: Giancarlo Di Giuseppe \email{digi.giancarlo@proton.me} (\href{https://orcid.org/0000-0001-5112-7560}{ORCID}) [copyright holder]

}
\keyword{internal}
