% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerR-package.R
\docType{package}
\name{cancerR-package}
\alias{cancerR}
\alias{cancerR-package}
\title{cancerR: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/giancarlodigi/cancerR}
  \item Report bugs at \url{https://github.com/giancarlodigi/cancerR/issues}
}

}
\author{
\strong{Maintainer}: Giancarlo Di Giuseppe \email{digi.giancarlo@proton.me} (\href{https://orcid.org/0000-0001-5112-7560}{ORCID}) [copyright holder]

}
\keyword{internal}
